Reformated files, deleted mainn and main
